Reimplemented Animations
Changed the order of sprites to be in a predictable order so that one set of offsets can play a pingpong animation given only a start frame.

Expanded animation metadata with frame timings and movement speed.

Adjusted the movement speed.

Fixed the "lurch" when remote players change how they're moving